Updated the code to look like the model answer & added the StartGame menu within a new package 'views'

+++ src/model/effects/Shock.java
@@ -4,23 +4,27 @@ import model.world.Champion;
 
 public class Shock extends Effect {
 
-    public Shock(int duration) {
-        super("Shock", duration, EffectType.DEBUFF);
-    }
+	public Shock(int duration) {
+		super("Shock", duration, EffectType.DEBUFF);
+		
+	}
 
-    @Override
-    public void apply(Champion c) {
-        c.setSpeed((int) (c.getSpeed() * 0.9));
-        c.setAttackDamage((int) (c.getAttackDamage() * 0.9));
-        c.setCurrentActionPoints((int) (c.getCurrentActionPoints() - 1));
-        c.setMaxActionPointsPerTurn((int) (c.getMaxActionPointsPerTurn() - 1));
-    }
+	@Override
+	public void apply(Champion c) {
+		c.setSpeed((int) (c.getSpeed()*0.9));
+		c.setAttackDamage((int) (c.getAttackDamage()*0.9));
+		c.setCurrentActionPoints(c.getCurrentActionPoints()-1);
+		c.setMaxActionPointsPerTurn(c.getMaxActionPointsPerTurn()-1);
+		
+	}
 
-    @Override
-    public void remove(Champion c) {
-        c.setSpeed((int) (c.getSpeed() / 0.9));
-        c.setAttackDamage((int) (c.getAttackDamage() / 0.9));
-        c.setCurrentActionPoints((int) (c.getCurrentActionPoints() + 1));
-        c.setMaxActionPointsPerTurn((int) (c.getMaxActionPointsPerTurn() + 1));
-    }
+	@Override
+	public void remove(Champion c) {
+		c.setSpeed((int) (c.getSpeed()/0.9));
+		c.setAttackDamage((int) (c.getAttackDamage()/0.9));
+		c.setCurrentActionPoints(c.getCurrentActionPoints()+1);
+		c.setMaxActionPointsPerTurn(c.getMaxActionPointsPerTurn()+1);
+		
+	}
+
 }
